Fix CI: gofmt drift, staticcheck TrimPrefix, extract pure helpers so Node tests run without SDK installed
Issue 1: Reformatted internal/advisor/advisor.go and internal/config/config.go with gofmt.

Issue 2: Replace strings.HasPrefix + manual slice in internal/overlay/overlay.go with strings.TrimPrefix.

Issue 3: Extract SDK-free helper functions into new modules:
- supervisor-helpers.mjs: computeStatusExtras, deriveRoleBehavior, makeUserMessage, handleSupervisorRequest (plus GIT_BRANCH_TTL_MS constant)
- socket-protocol.mjs: trySocketRequest, sockPathFor, sendAndFetchStatus, buildDeliveredEnvelope, buildErrorEnvelope, toolResult

This allows tests to import these pure functions without requiring @anthropic-ai/claude-agent-sdk to be installed. SDK-dependent tests (buildMessengerMcpServer) now gracefully skip when SDK is missing. All Node tests (39 pass, 7 skipped) and Go tests pass.
