+++ lib/agent-supervisor/sdk-mcp-tools.test.mjs
@@ -1,4 +1,4 @@
-import { test } from 'node:test'
+import { test, before } from 'node:test'
 import assert from 'node:assert/strict'
 import net from 'node:net'
 import fs from 'node:fs'
@@ -8,7 +8,7 @@ import {
   sendAndFetchStatus,
   buildDeliveredEnvelope,
   buildErrorEnvelope,
-} from './sdk-mcp-tools.mjs'
+} from './socket-protocol.mjs'
 
 function startFakeSupervisor(sockPath, statusReply) {
   const srv = net.createServer((conn) => {
@@ -87,8 +87,20 @@ test('buildErrorEnvelope shape', () => {
   assert.match(env.suggestion, /restart/)
 })
 
-test('worker role exposes notify_coordinator, ask_coordinator, shutdown_self', async () => {
-  const { buildMessengerMcpServer } = await import('./sdk-mcp-tools.mjs')
+let buildMessengerMcpServer = null
+let sdkAvailable = false
+
+before(async () => {
+  try {
+    ({ buildMessengerMcpServer } = await import('./sdk-mcp-tools.mjs'))
+    sdkAvailable = true
+  } catch (e) {
+    if (!e.message.includes('@anthropic-ai/claude-agent-sdk')) throw e
+    console.error('[sdk-mcp-tools.test.mjs] SDK not installed; skipping SDK-dependent tests')
+  }
+})
+
+test('worker role exposes notify_coordinator, ask_coordinator, shutdown_self', { skip: !sdkAvailable }, () => {
   const { toolNames } = buildMessengerMcpServer({
     role: 'agent',
     msgDir: '/tmp',
@@ -101,8 +113,7 @@ test('worker role exposes notify_coordinator, ask_coordinator, shutdown_self', a
   assert.ok(toolNames.includes('mcp__agent-messenger__shutdown_self'))
 })
 
-test('worker role exposes handshake_advisor and ask_advisor when advisors are configured', async () => {
-  const { buildMessengerMcpServer } = await import('./sdk-mcp-tools.mjs')
+test('worker role exposes handshake_advisor and ask_advisor when advisors are configured', { skip: !sdkAvailable }, () => {
   const { toolNames } = buildMessengerMcpServer({
     role: 'agent',
     msgDir: '/tmp',
@@ -114,8 +125,7 @@ test('worker role exposes handshake_advisor and ask_advisor when advisors are co
   assert.ok(toolNames.includes('mcp__agent-messenger__ask_advisor'))
 })
 
-test('coordinator role exposes ask_advisor and send_to_advisor when advisors are configured', async () => {
-  const { buildMessengerMcpServer } = await import('./sdk-mcp-tools.mjs')
+test('coordinator role exposes ask_advisor and send_to_advisor when advisors are configured', { skip: !sdkAvailable }, () => {
   const { toolNames } = buildMessengerMcpServer({
     role: 'coordinator',
     msgDir: '/tmp',
@@ -126,8 +136,7 @@ test('coordinator role exposes ask_advisor and send_to_advisor when advisors are
   assert.ok(toolNames.includes('mcp__agent-messenger__send_to_advisor'))
 })
 
-test('advisor tools are omitted when no advisors configured', async () => {
-  const { buildMessengerMcpServer } = await import('./sdk-mcp-tools.mjs')
+test('advisor tools are omitted when no advisors configured', { skip: !sdkAvailable }, () => {
   const { toolNames } = buildMessengerMcpServer({
     role: 'agent',
     msgDir: '/tmp',
@@ -139,8 +148,7 @@ test('advisor tools are omitted when no advisors configured', async () => {
   assert.ok(!toolNames.includes('mcp__agent-messenger__ask_advisor'))
 })
 
-test('coordinator role exposes send_to_worker', async () => {
-  const { buildMessengerMcpServer } = await import('./sdk-mcp-tools.mjs')
+test('coordinator role exposes send_to_worker', { skip: !sdkAvailable }, () => {
   const { toolNames } = buildMessengerMcpServer({
     role: 'coordinator',
     msgDir: '/tmp',
@@ -150,8 +158,7 @@ test('coordinator role exposes send_to_worker', async () => {
   assert.ok(toolNames.includes('mcp__agent-messenger__send_to_worker'))
 })
 
-test('advisor role exposes reply and note tools', async () => {
-  const { buildMessengerMcpServer } = await import('./sdk-mcp-tools.mjs')
+test('advisor role exposes reply and note tools', { skip: !sdkAvailable }, () => {
   const { toolNames } = buildMessengerMcpServer({
     role: 'advisor',
     msgDir: '/tmp',
