+++ lib/agent-supervisor/supervisor.integration.test.mjs
@@ -6,7 +6,7 @@ import path from 'node:path'
 import os from 'node:os'
 import {
   sendAndFetchStatus,
-} from './sdk-mcp-tools.mjs'
+} from './socket-protocol.mjs'
 
 const enabled = process.env.RUN_INTEGRATION === '1'
