+++ lib/agent-supervisor/supervisor.test.mjs
@@ -1,10 +1,9 @@
-import { test } from 'node:test'
+import { test, before } from 'node:test'
 import assert from 'node:assert/strict'
-// Import from args.mjs (the SDK-free module) so these tests can run
-// without node_modules/ installed. supervisor.mjs re-exports the same
-// names for runtime callers.
+// Import from SDK-free modules so these tests can run without node_modules/
+// installed. supervisor.mjs re-exports these for runtime callers.
 import { parseArgsForTest, buildQueryOptions } from './args.mjs'
-import { computeStatusExtras, deriveRoleBehavior } from './supervisor.mjs'
+import { computeStatusExtras, deriveRoleBehavior, handleSupervisorRequest } from './supervisor-helpers.mjs'
 
 test('parses --resume-session', () => {
   const args = parseArgsForTest(['node', 'supervisor.mjs',
@@ -153,7 +152,6 @@ test('computeStatusExtras caches git_branch within TTL', async () => {
 })
 
 test('shutdown_self closes the prompt queue', async () => {
-  const { handleSupervisorRequest } = await import('./supervisor.mjs')
   let shutdownCalled = false
   const state = {
     promptQueue: { push: () => {}, close: () => {}, _queue: [] },
